changes modified

+++ src/main/java/com/ii/testautomation/controllers/CompanyUserController.java
@@ -3,6 +3,7 @@ package com.ii.testautomation.controllers;
 import com.ii.testautomation.enums.RequestStatus;
 import com.ii.testautomation.response.common.BaseResponse;
 import com.ii.testautomation.service.CompanyUserService;
+import com.ii.testautomation.service.UserService;
 import com.ii.testautomation.utils.EndpointURI;
 import com.ii.testautomation.utils.StatusCodeBundle;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,9 +20,13 @@ public class CompanyUserController {
     private StatusCodeBundle statusCodeBundle;
     @Autowired
     private CompanyUserService companyUserService;
+    @Autowired
+    private UserService userService;
 
     @DeleteMapping(EndpointURI.COMPANY_USER_BY_ID)
     public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+        if (userService.existsByCompanyUserId(id))
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),statusCodeBundle.getCompanyUserDeleteDependentCode(),statusCodeBundle.getCompanyUserDeleteDependentMessage()));
         if (!companyUserService.existsById(id))
             return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),statusCodeBundle.getFailureCode(),statusCodeBundle.getCompanyUserIdNotExistMessage()));
         companyUserService.deleteById(id);
